+++ sdk/servicebus/azure-servicebus/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Breaking Changes**
 * Passing any type other than `ReceiveMode` as parameter `receive_mode` now throws a `TypeError` instead of `AttributeError`.
+* `AMQPMessage` (`Message.amqp_message`) properties are now read-only (will not be reflected in the underlying message) pending cross-language sdk design consensus.
 
 ## 7.0.0b6 (2020-09-10)